+++ src/sys/os/term/ansi/namespace/mod.rs
@@ -7,10 +7,12 @@
 mod tests;
 
 mod definition; // Ansi, control prefixes, erase escape codes
-mod cursor; // cursor escape codes
 mod terminal; // terminal modes
-mod font_mouse; // mouse and font effect escape codes
-mod osc; // AnsiLink
+mod cursor; // cursor escape codes
+mod mouse; // mouse escape codes
+mod font; // font effect escape codes
+mod color; // color escape codes
+mod osc; // AnsiLink, OSC codes
 
 crate::structural_mods! { // _mods
     _mods {